Replication is working. Fixed some very nasty bugs.
The trigger worked very weirdly. It prevented the modifications to
fresh (that was the problem even yesterday and not FEDERATED table
bug). Also, there was update bug in device table whick kept "changed"
in eternal change because column "fresh" was changed each sync.

There were also some typos in the source which copied some values
(like joindate and leavedate) as zeros.

Fixed the charset in tables, too.